Move entities into own package
