+++ src/main/java/com/devproserv/courses/model/Lecturer.java
@@ -1,4 +1,4 @@
-package com.devproserv.courses.dao;
+package com.devproserv.courses.model;
 
 /**
  * Represents the entity of the Lecturer. Maps the table 'lecturers' in the database.
